+++ src/Audit.WebApi/AuditApiAdapter.Core.cs
@@ -16,14 +16,12 @@ using System.IO;
 using System.Text;
 using Microsoft.AspNetCore.Mvc.Abstractions;
 using System.Reflection;
+using Microsoft.AspNetCore.Http.Extensions;
 
 namespace Audit.WebApi
 {
     internal class AuditApiAdapter
     {
-        private const string AuditApiActionKey = "__private_AuditApiAction__";
-        private const string AuditApiScopeKey = "__private_AuditApiScope__";
-        
         public bool IsActionIgnored(ActionExecutingContext actionContext)
         {
             var actionDescriptor = actionContext?.ActionDescriptor as ControllerActionDescriptor;
@@ -48,51 +46,60 @@ namespace Audit.WebApi
         {
             var httpContext = actionContext.HttpContext;
             var actionDescriptor = actionContext.ActionDescriptor as ControllerActionDescriptor;
-            var auditAction = new AuditApiAction
-            {
-                UserName = httpContext.User?.Identity.Name,
-                IpAddress = httpContext.Connection?.RemoteIpAddress?.ToString(),
-                RequestUrl = string.Format("{0}://{1}{2}", httpContext.Request.Scheme, httpContext.Request.Host, httpContext.Request.Path),
-                HttpMethod = actionContext.HttpContext.Request.Method,
-                FormVariables = GetFormVariables(httpContext),
-                Headers = includeHeaders ? ToDictionary(httpContext.Request.Headers) : null,
-                ActionName = actionDescriptor != null ? actionDescriptor.ActionName : actionContext.ActionDescriptor.DisplayName,
-                ControllerName = actionDescriptor?.ControllerName,
-                ActionParameters = GetActionParameters(actionDescriptor, actionContext.ActionArguments, serializeParams),
-                RequestBody = new BodyContent { Type = httpContext.Request.ContentType, Length = httpContext.Request.ContentLength, Value = includeRequestBody ? GetRequestBody(actionContext) : null },
-                TraceId = httpContext.TraceIdentifier
-            };
+
+            var auditAction = CreateOrUpdateAction(actionContext, includeHeaders, includeRequestBody, serializeParams, eventTypeName);
+
             var eventType = (eventTypeName ?? "{verb} {controller}/{action}").Replace("{verb}", auditAction.HttpMethod)
                 .Replace("{controller}", auditAction.ControllerName)
-                .Replace("{action}", auditAction.ActionName);
+                .Replace("{action}", auditAction.ActionName)
+                .Replace("{url}", auditAction.RequestUrl);
             // Create the audit scope
             var auditEventAction = new AuditEventWebApi()
             {
                 Action = auditAction
             };
             var auditScope = await AuditScope.CreateAsync(new AuditScopeOptions() { EventType = eventType, AuditEvent = auditEventAction, CallingMethod = actionDescriptor.MethodInfo });
-            httpContext.Items[AuditApiActionKey] = auditAction;
-            httpContext.Items[AuditApiScopeKey] = auditScope;
+            httpContext.Items[AuditApiHelper.AuditApiActionKey] = auditAction;
+            httpContext.Items[AuditApiHelper.AuditApiScopeKey] = auditScope;
         }
 
-        private IDictionary<string, string> GetFormVariables(HttpContext context)
+        private AuditApiAction CreateOrUpdateAction(ActionExecutingContext actionContext,
+            bool includeHeaders, bool includeRequestBody, bool serializeParams, string eventTypeName)
         {
-            if (!context.Request.HasFormContentType)
-            {
-                return null;
+            var httpContext = actionContext.HttpContext;
+            var actionDescriptor = actionContext.ActionDescriptor as ControllerActionDescriptor;
+            var action = httpContext.Items[AuditApiHelper.AuditApiActionKey] as AuditApiAction;
+            if (action == null)
+            { 
+                action = new AuditApiAction
+                {
+                    UserName = httpContext.User?.Identity.Name,
+                    IpAddress = httpContext.Connection?.RemoteIpAddress?.ToString(),
+                    HttpMethod = httpContext.Request.Method,
+                    FormVariables = AuditApiHelper.GetFormVariables(httpContext),
+                    TraceId = httpContext.TraceIdentifier
+                };
             }
-            IFormCollection formCollection;
-            try
+            action.RequestUrl = httpContext.Request.GetDisplayUrl();
+            action.ActionName = actionDescriptor != null ? actionDescriptor.ActionName : actionContext.ActionDescriptor.DisplayName;
+            action.ControllerName = actionDescriptor?.ControllerName;
+            action.ActionParameters = GetActionParameters(actionDescriptor, actionContext.ActionArguments, serializeParams);
+            if (includeHeaders)
             {
-                formCollection = context.Request.Form;
+                action.Headers = AuditApiHelper.ToDictionary(httpContext.Request.Headers);
             }
-            catch (InvalidDataException)
+            if (includeRequestBody)
             {
-                // InvalidDataException could be thrown if the form count exceeds the limit, etc
-                return null;
+                action.RequestBody = new BodyContent
+                {
+                    Type = httpContext.Request.ContentType,
+                    Length = httpContext.Request.ContentLength,
+                    Value = AuditApiHelper.GetRequestBody(httpContext)
+                };
             }
-            return ToDictionary(formCollection);
+            return action;
         }
+
 
         /// <summary>
         /// Occurs after the action method is invoked.
@@ -100,8 +107,8 @@ namespace Audit.WebApi
         public async Task AfterExecutedAsync(ActionExecutedContext context, bool includeModelState, bool includeResponseBody)
         {
             var httpContext = context.HttpContext;
-            var auditAction = httpContext.Items[AuditApiActionKey] as AuditApiAction;
-            var auditScope = httpContext.Items[AuditApiScopeKey] as AuditScope;
+            var auditAction = httpContext.Items[AuditApiHelper.AuditApiActionKey] as AuditApiAction;
+            var auditScope = httpContext.Items[AuditApiHelper.AuditApiScopeKey] as AuditScope;
             if (auditAction != null && auditScope != null)
             {
                 auditAction.Exception = context.Exception.GetExceptionInfo();
@@ -112,7 +119,7 @@ namespace Audit.WebApi
                     var statusCode = context.Result is ObjectResult && (context.Result as ObjectResult).StatusCode.HasValue ? (context.Result as ObjectResult).StatusCode.Value
                         : context.Result is StatusCodeResult ? (context.Result as StatusCodeResult).StatusCode : context.HttpContext.Response.StatusCode;
                     auditAction.ResponseStatusCode = statusCode;
-                    auditAction.ResponseStatus = GetStatusCodeString(auditAction.ResponseStatusCode);
+                    auditAction.ResponseStatus = AuditApiHelper.GetStatusCodeString(auditAction.ResponseStatusCode);
                     if (includeResponseBody)
                     {
                         var bodyType = context.Result?.GetType().GetFullTypeName();
@@ -127,9 +134,13 @@ namespace Audit.WebApi
                     auditAction.ResponseStatusCode = 500;
                     auditAction.ResponseStatus = "Internal Server Error";
                 }
-                // Replace the Action field and save
+                // Replace the Action field 
                 (auditScope.Event as AuditEventWebApi).Action = auditAction;
-                await auditScope.SaveAsync();
+                // Save, if action was not created by middleware
+                if (!auditAction.IsMiddleware)
+                {
+                    await auditScope.SaveAsync();
+                }
             }
         }
 
@@ -196,16 +207,6 @@ namespace Audit.WebApi
             return result.ToString();
         }
 
-        private string GetStatusCodeString(int statusCode)
-        {
-            var name = ((HttpStatusCode)statusCode).ToString();
-            string[] words = Regex.Matches(name, "(^[a-z]+|[A-Z]+(?![a-z])|[A-Z][a-z]+)")
-                .OfType<Match>()
-                .Select(m => m.Value)
-                .ToArray();
-            return words.Length == 0 ? name : string.Join(" ", words);
-        }
-
         private IDictionary<string, object> GetActionParameters(ControllerActionDescriptor actionDescriptor, IDictionary<string, object> actionArguments, bool serializeParams)
         {
             var args = actionArguments.ToDictionary(k => k.Key, v => v.Value); 
@@ -223,45 +224,9 @@ namespace Audit.WebApi
             return args;
         }
 
-        private static IDictionary<string, string> ToDictionary(IEnumerable<KeyValuePair<string, StringValues>> col)
-        {
-            if (col == null)
-            {
-                return null;
-            }
-            IDictionary<string, string> dict = new Dictionary<string, string>();
-            foreach (var k in col)
-            {
-                dict.Add(k.Key, string.Join(", ", k.Value));
-            }
-            return dict;
-        }
-
         internal static AuditScope GetCurrentScope(HttpContext httpContext)
         {
-            return httpContext.Items[AuditApiScopeKey] as AuditScope;
-        }
-
-        private string GetRequestBody(ActionExecutingContext actionContext)
-        {
-            var body = actionContext.HttpContext.Request.Body;
-            if (body != null && body.CanRead)
-            {
-                using (var stream = new MemoryStream())
-                {
-                    if (body.CanSeek)
-                    {
-                        body.Seek(0, SeekOrigin.Begin);
-                    }
-                    body.CopyTo(stream);
-                    if (body.CanSeek)
-                    {
-                        body.Seek(0, SeekOrigin.Begin);
-                    }
-                    return Encoding.UTF8.GetString(stream.ToArray());
-                }
-            }
-            return null;
+            return httpContext.Items[AuditApiHelper.AuditApiScopeKey] as AuditScope;
         }
 
     }