+++ test/Audit.Integration.AspNetCore/Program.cs
@@ -42,7 +42,39 @@ namespace Audit.Integration.AspNetCore
                 Console.WriteLine("START - TestInitialize");
                 await webApiTests.TestInitialize();
                 Console.WriteLine("PASSED - TestInitialize");
-                
+
+                Console.WriteLine("START - Test_WebApi_Post_Async");
+                await webApiTests.Test_WebApi_Post_Async();
+                Console.WriteLine("PASSED - Test_WebApi_Post_Async");
+
+                Console.WriteLine("START - Test_WebApi_DoubleActionFilter");
+                await webApiTests.Test_WebApi_DoubleActionFilter();
+                Console.WriteLine("PASSED - Test_WebApi_DoubleActionFilter");
+
+                Console.WriteLine("START - Test_WebApi_Middleware_NoResponseBody");
+                await webApiTests.Test_WebApi_Middleware_NoResponseBody();
+                Console.WriteLine("PASSED - Test_WebApi_Middleware_NoResponseBody");
+
+                Console.WriteLine("START - Test_WebApi_Middleware_WrongData");
+                await webApiTests.Test_WebApi_Middleware_WrongData();
+                Console.WriteLine("PASSED - Test_WebApi_Middleware_WrongData");
+
+                Console.WriteLine("START - Test_WebApi_Middleware_WrongRoute");
+                await webApiTests.Test_WebApi_Middleware_WrongRoute();
+                Console.WriteLine("PASSED - Test_WebApi_Middleware_WrongRoute");
+
+                Console.WriteLine("START - Test_WebApi_Middleware_Exception");
+                await webApiTests.Test_WebApi_Middleware_Exception();
+                Console.WriteLine("PASSED - Test_WebApi_Middleware_Exception");
+
+                Console.WriteLine("START - Test_WebApi_Middleware_Mix_Filter");
+                await webApiTests.Test_WebApi_Middleware_Mix_Filter();
+                Console.WriteLine("PASSED - Test_WebApi_Middleware_Mix_Filter");
+
+                Console.WriteLine("START - Test_WebApi_Middleware_Alone");
+                await webApiTests.Test_WebApi_Middleware_Alone();
+                Console.WriteLine("PASSED - Test_WebApi_Middleware_Alone");
+
                 Console.WriteLine("START - Test_WebApi_AuditIgnoreAttribute_Action_Async");
                 await webApiTests.Test_WebApi_AuditIgnoreAttribute_Action_Async();
                 Console.WriteLine("PASSED - Test_WebApi_AuditIgnoreAttribute_Action_Async");
@@ -62,10 +94,6 @@ namespace Audit.Integration.AspNetCore
                 Console.WriteLine("START - Test_WebApi_FilterResponseBody_Included");
                 await webApiTests.Test_WebApi_FilterResponseBody_Included();
                 Console.WriteLine("PASSED - Test_WebApi_FilterResponseBody_Included");
-
-                Console.WriteLine("START - Test_WebApi_Post_Async");
-                await webApiTests.Test_WebApi_Post_Async();
-                Console.WriteLine("PASSED - Test_WebApi_Post_Async");
 
                 Console.WriteLine("START - Test_Mvc_Exception_Async");
                 await mvcTests.Test_Mvc_Exception_Async();