+++ test/Audit.Integration.AspNetCore/WebApiTests.cs
@@ -26,6 +26,215 @@ namespace Audit.Integration.AspNetCore
             Assert.AreEqual("[\"value1\",\"value2\"]", s);
         }
 
+        public async Task Test_WebApi_Middleware_Mix_Filter()
+        {
+            // test using both mw and af
+            var insertEvs = new List<AuditEvent>();
+            Audit.Core.Configuration.Setup()
+                .UseDynamicProvider(_ => _.OnInsertAndReplace(ev =>
+                {
+                    insertEvs.Add(ev);
+                }))
+                .WithCreationPolicy(EventCreationPolicy.InsertOnEnd);
+
+            var url = $"http://localhost:{_port}/api/values/PostMix?middleware=123";
+            var client = new HttpClient();
+            var rqBody = "{\"value\": \"mix\"}";
+            var res = await client.PostAsync(url, new StringContent(rqBody, Encoding.UTF8, "application/json"));
+
+            Assert.AreEqual(HttpStatusCode.OK, res.StatusCode);
+            Assert.AreEqual(1, insertEvs.Count);
+            var action = insertEvs[0].GetWebApiAuditAction();
+            Assert.AreEqual("PostMix", action.ActionName);
+            Assert.AreEqual("Values", action.ControllerName);
+            Assert.AreEqual(true, action.IsMiddleware);
+            Assert.AreEqual("POST", action.HttpMethod);
+            Assert.AreEqual(2, action.ActionParameters.Count);
+            Assert.AreEqual(null, action.Exception);
+            Assert.AreEqual(rqBody, action.RequestBody.Value);
+            Assert.AreEqual(url, action.RequestUrl);
+            Assert.AreEqual("mix", action.ResponseBody.Value);
+            Assert.AreEqual(200, action.ResponseStatusCode);
+        }
+
+        public async Task Test_WebApi_Middleware_Exception()
+        {
+            // exception test using mw 
+            var insertEvs = new List<AuditEvent>();
+            Audit.Core.Configuration.Setup()
+                .UseDynamicProvider(_ => _.OnInsertAndReplace(ev =>
+                {
+                    insertEvs.Add(ev);
+                }))
+                .WithCreationPolicy(EventCreationPolicy.InsertOnEnd);
+
+            var url = $"http://localhost:{_port}/api/values/PostMiddleware?middleware=1";
+            var client = new HttpClient();
+            var rqBody = "{\"value\": \"666\"}";
+            var res = await client.PostAsync(url, new StringContent(rqBody, Encoding.UTF8, "application/json"));
+
+
+            Assert.AreEqual(HttpStatusCode.InternalServerError, res.StatusCode);
+            Assert.AreEqual(1, insertEvs.Count);
+            var action = insertEvs[0].GetWebApiAuditAction();
+            Assert.AreEqual(null, action.ActionName);
+            Assert.AreEqual(null, action.ControllerName);
+            Assert.AreEqual(true, action.IsMiddleware);
+            Assert.AreEqual("POST", action.HttpMethod);
+            Assert.AreEqual(null, action.ActionParameters);
+            Assert.IsNotNull(action.Exception);
+            Assert.IsTrue(action.Exception.ToUpper().Contains("THIS IS A TEST EXCEPTION 666"));
+            Assert.AreEqual(rqBody, action.RequestBody.Value);
+            Assert.AreEqual(url, action.RequestUrl);
+            Assert.AreEqual(null, action.ResponseBody);
+            Assert.AreEqual(500, action.ResponseStatusCode);
+        }
+
+        public async Task Test_WebApi_Middleware_WrongRoute()
+        {
+            var insertEvs = new List<AuditEvent>();
+            Audit.Core.Configuration.Setup()
+                .UseDynamicProvider(_ => _.OnInsertAndReplace(ev =>
+                {
+                    insertEvs.Add(ev);
+                }))
+                .WithCreationPolicy(EventCreationPolicy.InsertOnEnd);
+
+            var url = $"http://localhost:{_port}/api/values/doesnotexists?middleware=1";
+            var client = new HttpClient();
+            var rqBody = "{\"value\": \"x\"}";
+            var res = await client.PostAsync(url, new StringContent(rqBody, Encoding.UTF8, "application/json"));
+
+            Assert.AreEqual(HttpStatusCode.NotFound, res.StatusCode);
+            Assert.AreEqual(1, insertEvs.Count);
+            var action = insertEvs[0].GetWebApiAuditAction();
+            Assert.AreEqual(null, action.ActionName);
+            Assert.AreEqual(null, action.ControllerName);
+            Assert.AreEqual(true, action.IsMiddleware);
+            Assert.AreEqual("POST", action.HttpMethod);
+            Assert.AreEqual(null, action.ActionParameters);
+            Assert.IsNull(action.Exception);
+            Assert.AreEqual(rqBody, action.RequestBody.Value);
+            Assert.AreEqual(url, action.RequestUrl);
+            Assert.AreEqual(404, action.ResponseStatusCode);
+        }
+
+
+        public async Task Test_WebApi_Middleware_WrongData()
+        {
+            var insertEvs = new List<AuditEvent>();
+            Audit.Core.Configuration.Setup()
+                .UseDynamicProvider(_ => _.OnInsertAndReplace(ev =>
+                {
+                    insertEvs.Add(ev);
+                }))
+                .WithCreationPolicy(EventCreationPolicy.InsertOnEnd);
+
+            var url = $"http://localhost:{_port}/api/values/PostMiddleware?middleware=1";
+            var client = new HttpClient();
+            var rqBody = "{\"value\": \"123\"}";
+            var res = await client.PostAsync(url, new StringContent(rqBody, Encoding.UTF8, "unknown/unknown"));
+
+            Assert.AreEqual(HttpStatusCode.UnsupportedMediaType, res.StatusCode);
+            Assert.AreEqual(1, insertEvs.Count);
+            var action = insertEvs[0].GetWebApiAuditAction();
+            Assert.AreEqual(null, action.ActionName);
+            Assert.AreEqual(null, action.ControllerName);
+            Assert.AreEqual(true, action.IsMiddleware);
+            Assert.AreEqual("POST", action.HttpMethod);
+            Assert.AreEqual(null, action.ActionParameters);
+            Assert.IsNull(action.Exception);
+            Assert.AreEqual(rqBody, action.RequestBody.Value);
+            Assert.AreEqual(url, action.RequestUrl);
+            Assert.AreEqual(415, action.ResponseStatusCode);
+        }
+
+        public async Task Test_WebApi_Middleware_NoResponseBody()
+        {
+            var insertEvs = new List<AuditEvent>();
+            Audit.Core.Configuration.Setup()
+                .UseDynamicProvider(_ => _.OnInsertAndReplace(ev =>
+                {
+                    insertEvs.Add(ev);
+                }))
+                .WithCreationPolicy(EventCreationPolicy.InsertOnEnd);
+
+            var url = $"http://localhost:{_port}/api/values/TestIgnoreAction?middleware=123&noresponsebody=1";
+            var client = new HttpClient();
+            var rqBody = "{\"value\": \"def\"}";
+            var res = await client.PostAsync(url, new StringContent(rqBody, Encoding.UTF8, "application/json"));
+
+            Assert.AreEqual(HttpStatusCode.OK, res.StatusCode);
+            Assert.AreEqual(1, insertEvs.Count);
+            var action = insertEvs[0].GetWebApiAuditAction();
+            Assert.AreEqual(null, action.ActionName);
+            Assert.AreEqual(null, action.ControllerName);
+            Assert.AreEqual(true, action.IsMiddleware);
+            Assert.AreEqual("POST", action.HttpMethod);
+            Assert.AreEqual(null, action.Exception);
+            Assert.AreEqual(rqBody, action.RequestBody.Value);
+            Assert.AreEqual(url, action.RequestUrl);
+            Assert.AreEqual(null, action.ResponseBody);
+            Assert.AreEqual(200, action.ResponseStatusCode);
+        }
+
+        public async Task Test_WebApi_DoubleActionFilter()
+        {
+            var insertEvs = new List<AuditEvent>();
+            Audit.Core.Configuration.Setup()
+                .UseDynamicProvider(_ => _.OnInsertAndReplace(ev =>
+                {
+                    insertEvs.Add(ev);
+                }))
+                .WithCreationPolicy(EventCreationPolicy.InsertOnEnd);
+
+            var url = $"http://localhost:{_port}/api/MoreValues?middleware=123";
+            var client = new HttpClient();
+            var res = await client.GetAsync(url);
+
+            Assert.AreEqual(HttpStatusCode.OK, res.StatusCode);
+            Assert.AreEqual(1, insertEvs.Count);
+            var action = insertEvs[0].GetWebApiAuditAction();
+            Assert.AreEqual("Get", action.ActionName);
+            Assert.AreEqual("MoreValues", action.ControllerName);
+            Assert.AreEqual(true, action.IsMiddleware);
+            Assert.AreEqual("GET", action.HttpMethod);
+            Assert.AreEqual(null, action.Exception);
+            Assert.AreEqual(url, action.RequestUrl);
+            Assert.IsNotNull(action.ResponseBody);
+            Assert.AreEqual(200, action.ResponseStatusCode);
+
+        }
+
+        public async Task Test_WebApi_Middleware_Alone()
+        {
+            var insertEvs = new List<AuditEvent>();
+            Audit.Core.Configuration.Setup()
+                .UseDynamicProvider(_ => _.OnInsert(ev =>
+                {
+                    insertEvs.Add(ev);
+                    return Guid.NewGuid();
+                }))
+                .WithCreationPolicy(EventCreationPolicy.InsertOnEnd);
+
+            
+            var url = $"http://localhost:{_port}/api/values/TestIgnoreAction?middleware=123";
+            var client = new HttpClient();
+            var res = await client.PostAsync(url, new StringContent("{\"value\": \"def\"}", Encoding.UTF8, "application/json"));
+
+            Assert.AreEqual(HttpStatusCode.OK, res.StatusCode);
+            Assert.AreEqual(1, insertEvs.Count);
+            var action = insertEvs[0].GetWebApiAuditAction();
+            Assert.IsTrue(action.IsMiddleware);
+            Assert.AreEqual("POST", action.HttpMethod);
+            Assert.IsNull(action.ActionName);
+            Assert.IsNull(action.ControllerName);
+            Assert.IsTrue(action.Headers.Count > 0);
+            Assert.AreEqual("{\"value\": \"def\"}", action.RequestBody.Value.ToString());
+            Assert.AreEqual("hi", action.ResponseBody.Value.ToString());
+            Assert.AreEqual(url, action.RequestUrl);
+        }
+
         public async Task Test_WebApi_AuditIgnoreAttribute_Action_Async()
         {
             var insertEvs = new List<AuditEvent>();
@@ -70,6 +279,8 @@ namespace Audit.Integration.AspNetCore
             Assert.AreEqual(1, insertEvs[0].GetWebApiAuditAction().ActionParameters.Count);
             Assert.IsTrue(insertEvs[0].GetWebApiAuditAction().ActionParameters.ContainsKey("user"));
             Assert.IsFalse(insertEvs[0].GetWebApiAuditAction().ActionParameters.ContainsKey("password"));
+            Assert.IsFalse(insertEvs[0].GetWebApiAuditAction().IsMiddleware);
+            Assert.AreEqual(url, insertEvs[0].GetWebApiAuditAction().RequestUrl);
         }
 
         public async Task Test_WebApi_FormCollectionLimit_Async()
@@ -136,6 +347,7 @@ namespace Audit.Integration.AspNetCore
             Assert.AreEqual("def", insertEvs[0].GetWebApiAuditAction().ResponseBody.Value);
             Assert.AreEqual(200, insertEvs[0].GetWebApiAuditAction().ResponseStatusCode);
             Assert.AreEqual("POST.Values.GlobalAudit", insertEvs[0].EventType);
+            Assert.IsFalse(insertEvs[0].GetWebApiAuditAction().IsMiddleware);
         }
 
         public async Task Test_WebApi_HappyPath_Async()
@@ -163,6 +375,7 @@ namespace Audit.Integration.AspNetCore
             Assert.AreEqual(1, replaceEvs.Count);
             Assert.AreEqual(null, insertEvs[0].ResponseBody);
             Assert.AreEqual("10", replaceEvs[0].ResponseBody.Value);
+            Assert.IsFalse(insertEvs[0].IsMiddleware);
         }
 
         public async Task Test_WebApi_Post_Async()